refactor: moving core sanity checks to cfg.ValidateBasic and populating default conf
